+++ app/controllers/tabs/passwords.py
@@ -2,27 +2,28 @@ import logging
 import traceback
 import typing
 import uuid
-from datetime import datetime
 from enum import StrEnum
 from functools import partial
 
 from pydantic import AnyUrl, TypeAdapter, ValidationError
-from PySide6.QtCore import QAbstractTableModel, QModelIndex, QObject, Qt, Signal, Slot
+from PySide6.QtCore import QModelIndex, QObject, Qt, Signal, Slot
 from PySide6.QtGui import QAction, QIcon, QStandardItem, QStandardItemModel
 from PySide6.QtWidgets import QDialog, QDialogButtonBox, QHeaderView, QMenu, QMessageBox
 
 from ...localdb.database import MainDatabase
-from ...models import (
+from ...models.models import (
     AddPasswordGroup,
     EditedEntryWithGroup,
     EditedPasswordEntryInfo,
     GroupParentData,
     PasswordEntryData,
 )
+from ...models.ui import PasswordEntriesTableModel
 from ...serversync.client import SyncClient
-from ...serversync.models import EntryPublicGet, GenericSuccess, GroupPublicModify
+from ...serversync.models import GenericSuccess, GroupPublicModify
 from ...ui.add_password_group_dialog import Ui_AddPasswordGroupDialog
 from ...ui.password_entry_info_dialog import Ui_PasswordEntryInfoDialog
+from ..entry_data import EntriesDataController
 from ...workers import make_worker_thread
 
 if typing.TYPE_CHECKING:
@@ -38,107 +39,6 @@ class EmitDialogInfoAs(StrEnum):
     edit = "edit"
 
 
-class PasswordEntriesTableModel(QAbstractTableModel):
-    def __init__(self, /, parent: "PasswordEntriesController" = None):
-        super().__init__(parent)
-        self.pw_ctrl = parent
-
-        self._display_data: list[list[str | datetime]] = []
-        self._item_data: list[PasswordEntryData] = []
-
-        self._col_headers: list[str] = ["Title", "Username", "URL", "Created At"]
-
-    def rowCount(self, /, parent: QModelIndex):
-        return len(self._item_data)
-
-    def columnCount(self, /, parent: QModelIndex):
-        return len(self._col_headers)
-
-    def data(self, index, /, role):
-        if role == Qt.ItemDataRole.UserRole:
-            return self._item_data[index.row()]
-
-        if role == Qt.ItemDataRole.DisplayRole:
-            value = self._display_data[index.row()][index.column()]
-            if isinstance(value, datetime):
-                return value.strftime("%x %X")
-
-            if isinstance(value, AnyUrl):
-                return str(value)
-
-            if value is None:
-                logger.debug("Received 'None' value at cell [%d, %d]", index.row(), index.column())
-                return ""
-
-            return value
-
-        return None
-
-    def headerData(self, section, orientation, /, role):
-        if role != Qt.ItemDataRole.DisplayRole:
-            return None
-
-        if orientation == Qt.Orientation.Horizontal:
-            return self._col_headers[section]
-
-        if orientation == Qt.Orientation.Vertical:
-            return super().headerData(section, orientation, role)
-
-    def load_entries(self, entries: list[PasswordEntryData]):
-        self.beginResetModel()
-
-        self._display_data.clear()
-        self._item_data.clear()
-
-        logger.debug("Cleared all model entries")
-
-        for entry in entries:
-            display_data = [entry.title, entry.username, entry.url, entry.created_at]
-            self._display_data.append(display_data)
-
-            self._item_data.append(entry)
-
-        logger.debug("Added %d entries to model", len(self._item_data))
-        self.endResetModel()
-
-    def add_entry(self, entry: PasswordEntryData):
-        display_data = [entry.title, entry.username, entry.url, entry.created_at]
-
-        self._item_data.append(entry)
-        self._display_data.append(display_data)
-
-        self.layoutChanged.emit()
-
-    def delete_entry(self, index: QModelIndex):
-        if not index.isValid():
-            return
-
-        del self._item_data[index.row()]
-        del self._display_data[index.row()]
-
-        self.layoutChanged.emit()
-
-    def update_entry(self, data: PasswordEntryData):
-        entries = [item for item in self._item_data if data.entry_id == item.entry_id]
-        if not entries:
-            logger.warning("Entry ID '%s' not in model", data.entry_id)
-            return
-
-        entry = entries[0]
-        row = self._item_data.index(entry)
-
-        display_data = [data.title, data.username, data.url, data.created_at]
-
-        self._item_data[row] = data
-        self._display_data[row] = display_data
-
-        top = self.index(row, 0)
-        bottom = self.index(row, self.columnCount() - 1)
-
-        self.dataChanged.emit(top, bottom)
-
-
-# TODO: Make this ideally wait for the client to either load or be disabled
 class PasswordsTabController(QObject):
     useClient = Signal(SyncClient)
 
@@ -151,37 +51,32 @@ class PasswordsTabController(QObject):
         self.ui = self.mw_parent.ui
         self.db: MainDatabase = None
 
-        self.client: SyncClient | None = None
+        self.client: SyncClient = None
 
     @Slot(MainDatabase)
     def database_loaded(self, database: MainDatabase):
         self.db = database
-        make_worker_thread(self.db.groups.get_children_of_root, self.setup, self.worker_exc_received)
-        self.ui.statusbar.showMessage("Passwords - Loading top-level entries", timeout=5000)
+        logger.debug("Local database loaded, waiting for client")
 
     @Slot(SyncClient)
     def client_loaded(self, client: SyncClient):
+        if not self.db:
+            raise RuntimeError("client loaded before database")
+
         self.client = client
-        self.useClient.emit(self.client)
+        logger.debug("Client loaded, setting up controllers")
 
-        logger.debug("Sent signal to use sync client")
+        self.setup()
 
-    @Slot(GroupParentData)
-    def setup(self, data: GroupParentData):
-        self.current_group = data
-
+    def setup(self):
         # Controllers
-        self.entries_ctrl = PasswordEntriesController(self)
-        self.groups_ctrl = PasswordGroupsItemController(self)
+        self.groups_ctrl = PasswordGroupsItemController(self, self.db, self.client)
+        self.entries_ctrl = PasswordEntriesController(self, self.db, self.client)
 
         self.entry_info_ctrl = PasswordEntryInfoController(self)
         self.groups_ctrl.groupChanged.connect(self.entries_ctrl.reload_entries)
 
         self.entries_ctrl.entryChanged.connect(self.entry_info_ctrl.entry_changed)
-        self.useClient.connect(self.entries_ctrl.client_loaded)
-
-        self.useClient.connect(self.groups_ctrl.client_loaded)
-        self.ui.statusbar.showMessage("Passwords - Top-level entries loaded", timeout=5000)
 
     @Slot(Exception)
     def worker_exc_received(self, exc: Exception):
@@ -201,14 +96,16 @@ class PasswordsTabController(QObject):
 class PasswordEntriesController(QObject):
     entryChanged = Signal(PasswordEntryData)
 
-    def __init__(self, pw_parent: PasswordsTabController):
+    def __init__(self, pw_parent: PasswordsTabController, database: MainDatabase, client: SyncClient):
         super().__init__(pw_parent)
 
         self.mw_parent = pw_parent.mw_parent
         self.pw_parent = pw_parent
 
         self.ui = self.mw_parent.ui
-        self.db = pw_parent.db
+
+        self.db: MainDatabase = database
+        self.client: SyncClient = client
 
         self.entries_model = PasswordEntriesTableModel(parent=self)
         self.ui.passwordEntriesTableView.setModel(self.entries_model)
@@ -216,7 +113,6 @@ class PasswordEntriesController(QObject):
         self.entry_info_dialog = PasswordEntryInfoDialog(self)
         self.data_ctrl: EntriesDataController = EntriesDataController(self)
 
-        self.client: SyncClient | None = None
         self.setup()
 
     def setup(self):
@@ -229,39 +125,16 @@ class PasswordEntriesController(QObject):
         self.ui.passwordEntriesTableView.customContextMenuRequested.connect(self.context_menu_event)
 
         self.ui.passwordEntriesTableView.clicked.connect(self.tableview_clicked)
-        self.entry_info_dialog.addEntryRequested.connect(self.data_ctrl.add_entry_dialog_accepted)
+        self.ui.passwordEntriesTableView.doubleClicked.connect(self.tableview_doubleclicked)
 
-        self.entry_info_dialog.editEntryRequested.connect(self.data_ctrl.edit_entry_dialog_accepted)
+        self.entry_info_dialog.addEntryRequested.connect(self.data_ctrl.add_entry.start_processing)
+        self.entry_info_dialog.editEntryRequested.connect(self.data_ctrl.update_entry.start_processing)
 
-        self.data_ctrl.addEntryComplete.connect(self.model_add_password_entry)
-        self.data_ctrl.deleteEntryComplete.connect(self.model_delete_password_entry)
+        self.data_ctrl.add_entry.addEntryComplete.connect(self.model_add_password_entry)
+        self.data_ctrl.delete_entry.deleteEntryComplete.connect(self.model_delete_password_entry)
 
-        self.data_ctrl.editEntryComplete.connect(self.model_edit_password_entry)
-        self.data_ctrl.entriesReloaded.connect(self.model_reload_entries)
-
-    @Slot(SyncClient)
-    def client_loaded(self, client: SyncClient):
-        self.client = client
-        self.data_ctrl.client_loaded(client)
-
-    @Slot(QModelIndex)
-    def tableview_clicked(self, index: QModelIndex):
-        data: PasswordEntryData = index.data(Qt.ItemDataRole.UserRole)
-        self.entryChanged.emit(data)
-
-    @Slot(GroupParentData)
-    def reload_entries(self, group: GroupParentData):
-        self.current_group = group
-        self.data_ctrl.group_changed(group)
-
-        self.ui.statusbar.showMessage(f"Passwords - Reloading entries for group '{group.group_name}'", timeout=5000)
-
-    @Slot(list)
-    def model_reload_entries(self, entries: list[PasswordEntryData]):
-        logger.info("Fetched %d entries", len(entries))
-        self.entries_model.load_entries(entries)
-
-        self.ui.statusbar.showMessage("Passwords - Entries reloaded", timeout=5000)
+        self.data_ctrl.update_entry.updateEntryComplete.connect(self.model_edit_password_entry)
+        self.data_ctrl.fetch_entries.fetchEntriesComplete.connect(self.model_reload_entries)
 
     @Slot()
     def context_menu_event(self, pos):
@@ -290,8 +163,8 @@ class PasswordEntriesController(QObject):
             index = indexes[0]
             current_item: PasswordEntryData = self.entries_model.data(index, Qt.ItemDataRole.UserRole)
 
-            remove_entry_action.triggered.connect(lambda: self.delete_password_entry(index, current_item))
-            edit_entry_action.triggered.connect(lambda: self.edit_password_entry(index, current_item))
+            remove_entry_action.triggered.connect(lambda: self.delete_password_entry(current_item))
+            edit_entry_action.triggered.connect(lambda: self.edit_password_entry(current_item))
 
         # Item-dependent actions
         if indexes:
@@ -299,6 +172,30 @@ class PasswordEntriesController(QObject):
             context.addAction(edit_entry_action)
 
         context.exec(self.ui.passwordEntriesTableView.mapToGlobal(pos))
+
+    @Slot(QModelIndex)
+    def tableview_clicked(self, index: QModelIndex):
+        data: PasswordEntryData = index.data(Qt.ItemDataRole.UserRole)
+        self.entryChanged.emit(data)
+
+    @Slot(QModelIndex)
+    def tableview_doubleclicked(self, index: QModelIndex):
+        data: PasswordEntryData = index.data(Qt.ItemDataRole.UserRole)
+        self.edit_password_entry(data)
+
+    @Slot(GroupParentData)
+    def reload_entries(self, group: GroupParentData):
+        self.data_ctrl.add_entry.group_changed(group)
+        self.data_ctrl.fetch_entries.start_processing(group)
+
+        self.ui.statusbar.showMessage(f"Passwords - Reloading entries for group '{group.group_name}'", timeout=5000)
+
+    @Slot(list)
+    def model_reload_entries(self, entries: list[PasswordEntryData]):
+        logger.info("Fetched %d entries", len(entries))
+        self.entries_model.load_entries(entries)
+
+        self.ui.statusbar.showMessage("Passwords - Entries reloaded", timeout=5000)
 
     @Slot()
     def add_password_entry(self):
@@ -312,7 +209,7 @@ class PasswordEntriesController(QObject):
 
         self.entries_model.add_entry(entry)
 
-    def delete_password_entry(self, index: QModelIndex, item: PasswordEntryData):
+    def delete_password_entry(self, item: PasswordEntryData):
         btn = QMessageBox.information(
             self.mw_parent,
             "PasswordManager - Client",
@@ -326,16 +223,16 @@ class PasswordEntriesController(QObject):
         logger.info("Deleting entry '%s'", item.title)
         self.ui.statusbar.showMessage(f"Passwords - Deleting entry '{item.title}'", timeout=5000)
 
-        self.entries_model.delete_entry(index)
+        self.data_ctrl.delete_entry.start_processing(item)
+
+    @Slot(PasswordEntryData)
+    def model_delete_password_entry(self, entry: PasswordEntryData):
+        self.entries_model.delete_entry(entry)
         self.ui.passwordEntriesTableView.clearSelection()
 
-        self.data_ctrl.delete_password_entry(item)
-
-    @Slot(bool)
-    def model_delete_password_entry(self, success: bool):
         self.ui.statusbar.showMessage("Passwords - Entry deleted")
 
-    def edit_password_entry(self, index: QModelIndex, item: PasswordEntryData):
+    def edit_password_entry(self, item: PasswordEntryData):
         self.entry_info_dialog.reset_data(emit_as=EmitDialogInfoAs.edit)
         self.entry_info_dialog.set_existing_data(item)
 
@@ -350,6 +247,7 @@ class PasswordEntriesController(QObject):
 
         indexes = self.ui.passwordEntriesTableView.selectedIndexes()
         if not indexes:
+            logger.debug("No selected indexes")
             return
 
         index = indexes[0]
@@ -357,39 +255,39 @@ class PasswordEntriesController(QObject):
             logger.debug("Entry changed is the currently selected entry, updating")
             self.entryChanged.emit(entry)
 
+        return
+
 
 # TODO: Ensure when setting up sync for the first time, root group always matches the server
+# TODO: Use an actual tree model instead of relying on an item model
 class PasswordGroupsItemController(QObject):
     groupChanged = Signal(GroupParentData)
 
-    def __init__(self, pw_parent: PasswordsTabController):
+    def __init__(self, pw_parent: PasswordsTabController, database: MainDatabase, client: SyncClient):
         super().__init__(pw_parent)
 
         self.mw_parent = pw_parent.mw_parent
         self.pw_parent = pw_parent
 
         self.ui = self.mw_parent.ui
-        self.db = pw_parent.db
 
-        self.add_group_dialog = AddPasswordGroupDialog(self)
+        self.db: MainDatabase = database
+        self.client: SyncClient = client
+
         self.worker_exc_received = pw_parent.worker_exc_received
 
+        # Controllers here
+        self.add_group_dialog = AddPasswordGroupDialog(self)
         self.groups_model = QStandardItemModel(parent=self)
-        self.current_group: GroupParentData = None
 
+        self.current_group: GroupParentData = None
         self.data_ctrl: GroupsDataController = GroupsDataController(self)
-        self.client: SyncClient | None = None
 
         # TODO: Refactor this if needed
         self.root_item: QStandardItem = None
         self._items: dict[uuid.UUID, QStandardItem] = {}
 
         self.setup()
-
-    @Slot(SyncClient)
-    def client_loaded(self, client: SyncClient):
-        self.client = client
-        self.data_ctrl.client_loaded(client)
 
     def setup(self):
         self.ui.passwordGroupsTreeView.setModel(self.groups_model)
@@ -454,18 +352,18 @@ class PasswordGroupsItemController(QObject):
 
     def load_groups(self, parent_group: GroupParentData, parent_item: QStandardItem | None = None):
         """Queries the database, use this in a worker thread."""
-        if self.client:
+        if self.client.enabled:
             net_children = self.client.groups.get_children_of_group(parent_group.group_id)
             db_children = self.db.groups.get_children_of_group(parent_group.group_id)
 
-            len_net_children = len(net_children)
-            len_db_children = len(db_children)
+            len_net_children = len(net_children.child_groups)
+            len_db_children = len(db_children.child_groups)
+
             if len_net_children != len_db_children:
                 logger.error("Expected %d groups locally, instead got %d groups", len_net_children, len_db_children)
-                raise ValueError("Server groups do not match local groups")
+                # raise ValueError("Server groups do not match local groups")
 
-            ta = TypeAdapter(list[GroupParentData])
-            children = ta.validate_python(net_children)
+            children = GroupParentData.model_validate(net_children, from_attributes=True)
         else:
             children = self.db.groups.get_children_of_group(parent_group.group_id)
 
@@ -476,7 +374,7 @@ class PasswordGroupsItemController(QObject):
             item = self.append_group(group, parentItem)
             self._items[group.group_id] = item
 
-            logger.info(
+            logger.debug(
                 "Got child group [%s - %s] of group '%s'", group.group_id, group.group_name, parent_group.group_name
             )
             self.load_groups(group, parent_item=item)
@@ -490,7 +388,7 @@ class PasswordGroupsItemController(QObject):
         item.setData(group, role=Qt.ItemDataRole.UserRole)
         parentItem.appendRow(item)
 
-        logger.info(
+        logger.debug(
             "Appended group [%s - %s] to tree model",
             group.group_id,
             group.group_name,
@@ -616,7 +514,7 @@ class PasswordEntryInfoController(QObject):
         self._entry: PasswordEntryData = None
         self._hidden: bool = True
 
-    @Slot()
+    @Slot(PasswordEntryData)
     def entry_changed(self, entry: PasswordEntryData):
         logger.debug("Entry changed to '%s', changing info", entry.title)
         self._entry = entry
@@ -754,135 +652,6 @@ class AddPasswordGroupDialog(QDialog):
             btn.setEnabled(True)
 
 
-# TODO: Refactor to make it so that istead of storing all CRUD opts in one class,
-# its split into many classes with either add, delete, edit or read
-class EntriesDataController(QObject):
-    addEntryComplete = Signal(PasswordEntryData)
-    deleteEntryComplete = Signal(bool)
-
-    editEntryComplete = Signal(PasswordEntryData)
-    entriesReloaded = Signal(list)  # list[PasswordEntryData]
-
-    def __init__(self, parent: PasswordEntriesController):
-        super().__init__(parent)
-
-        self.mw_parent = parent.mw_parent
-        self.ctrl_parent = parent
-
-        self.ui = self.mw_parent.ui
-
-        self.db: MainDatabase = parent.db
-        self.client: SyncClient | None = None
-
-        self.worker_exc_received = self.ctrl_parent.pw_parent.worker_exc_received
-
-    def client_loaded(self, client: SyncClient):
-        self.client = client
-
-    @Slot(GroupParentData)
-    def group_changed(self, group: GroupParentData):
-        self.current_group = group
-        func = partial(self.db.entries.get_entries_by_group, group.group_id)
-
-        make_worker_thread(func, self.entries_reloaded, self.worker_exc_received)
-        logger.info("Reloading entries for group '%s'", group.group_name)
-
-    @Slot(list)
-    def entries_reloaded(self, entries: list[PasswordEntryData]):
-        self.entriesReloaded.emit(entries)
-
-    @Slot(EditedPasswordEntryInfo)
-    def add_entry_dialog_accepted(self, data: EditedPasswordEntryInfo):
-        if self.client:
-            net_func = partial(self.client.entries.create_entry, self.current_group.group_id, data)
-            make_worker_thread(net_func, self.sync_add_password_entry, self.worker_exc_received)
-
-            logger.info("Sending request to add entry...")
-            self.ui.statusbar.showMessage("Passwords - Sent request to add password entry", timeout=5000)
-
-            return
-
-        func = partial(self.db.entries.create_entry, self.current_group.group_id, data)
-        make_worker_thread(func, self.add_complete, self.worker_exc_received)
-
-    @Slot(EntryPublicGet)
-    def sync_add_password_entry(self, entry: EntryPublicGet):
-        url_or_none = entry.url.model_dump() if entry.url is not None else None
-
-        data = EditedPasswordEntryInfo(
-            title=entry.title, username=entry.username, password=entry.password, url=url_or_none, notes=entry.notes
-        )
-        func = partial(
-            self.db.entries.create_entry, entry.group_id, data, entry_id=entry.entry_id, created_at=entry.created_at
-        )
-        make_worker_thread(func, self.add_complete, self.worker_exc_received)
-        logger.info("Adding entry to local database")
-
-    @Slot(PasswordEntryData)
-    def add_complete(self, entry: PasswordEntryData):
-        self.addEntryComplete.emit(entry)
-
-    def delete_password_entry(self, item: PasswordEntryData):
-        @Slot(GenericSuccess)
-        def sync_delete_password_entry(data: GenericSuccess):
-            func = partial(self.db.entries.delete_entry_by_id, item.entry_id, item.group_id)
-            make_worker_thread(func, model_delete_password_entry, self.worker_exc_received)
-
-        @Slot(GenericSuccess)
-        def model_delete_password_entry(success: bool):
-            self.deleteEntryComplete.emit(success)
-
-        if self.client:
-            net_func = partial(self.client.entries.delete_entry_by_id, item.entry_id, item.group_id)
-            make_worker_thread(net_func, sync_delete_password_entry, self.worker_exc_received)
-
-            logger.info("Sent request to delete entry '%s'", item.title)
-            return
-
-        func = partial(self.db.entries.delete_entry_by_id, item.entry_id, item.group_id)
-        make_worker_thread(func, model_delete_password_entry, self.worker_exc_received)
-
-    @Slot(EditedEntryWithGroup)
-    def edit_entry_dialog_accepted(self, data: EditedEntryWithGroup):
-        if self.client:
-            net_func = partial(self.client.entries.update_entry_data, data.entry_id, data)
-            make_worker_thread(net_func, self.sync_edit_password_entry, self.worker_exc_received)
-
-            logger.info("Sending request to edit entry...")
-            self.ui.statusbar.showMessage("Passwords - Sent request to edit password entry", timeout=5000)
-
-            return
-
-        func = partial(self.db.entries.update_entry_data, data.entry_id, data)
-        make_worker_thread(func, self.edit_complete, self.worker_exc_received)
-
-    @Slot(EntryPublicGet)
-    def sync_edit_password_entry(self, data: EntryPublicGet):
-        # This is used because it's a RootModel generated instead of an actual AnyUrl
-        url_or_none = data.url.model_dump() if data.url is not None else None
-
-        entry = EditedEntryWithGroup(
-            title=data.title,
-            username=data.username,
-            password=data.password,
-            url=url_or_none,
-            notes=data.notes,
-            entry_id=data.entry_id,
-            group_id=data.group_id,
-        )
-        func = partial(self.db.entries.update_entry_data, entry.entry_id, entry)
-        make_worker_thread(func, self.model_edit_password_entry, self.worker_exc_received)
-
-    @Slot(EntryPublicGet)
-    def model_edit_password_entry(self, data: EntryPublicGet):
-        func = partial(self.db.entries.update_entry_data, data.entry_id, data)
-        make_worker_thread(func, self.edit_complete, self.worker_exc_received)
-
-    @Slot(PasswordEntryData)
-    def edit_complete(self, data: PasswordEntryData):
-        self.editEntryComplete.emit(data)
-
-
 class GroupsDataController(QObject):
     addGroupComplete = Signal(GroupParentData)
     # deleteGroupComplete = Signal(bool)
@@ -896,13 +665,10 @@ class GroupsDataController(QObject):
         self.ui = self.mw_parent.ui
 
         self.db: MainDatabase = parent.db
-        self.client: SyncClient | None = None
+        self.client: SyncClient = parent.client
 
         self.current_group: GroupParentData = None
         self.worker_exc_received = self.ctrl_parent.pw_parent.worker_exc_received
-
-    def client_loaded(self, client: SyncClient):
-        self.client = client
 
     def group_changed(self, group: GroupParentData):
         self.current_group = group