+++ app/models/ui.py
@@ -1,0 +1,157 @@
+import logging
+import typing
+import uuid
+from datetime import datetime
+from pathlib import Path
+
+from pydantic import AnyUrl
+from PySide6.QtCore import QAbstractListModel, QAbstractTableModel, QModelIndex, Qt
+
+from .models import PasswordEntryData
+
+if typing.TYPE_CHECKING:
+    from ..controllers.tabs.databases import DatabasesTabController
+    from ..controllers.tabs.passwords import PasswordEntriesController
+
+logger: logging.Logger = logging.getLogger("passwordmanager-client")
+
+
+class PasswordEntriesTableModel(QAbstractTableModel):
+    def __init__(self, /, parent: "PasswordEntriesController" = None):
+        super().__init__(parent)
+        self.pw_ctrl = parent
+
+        self._display_data: list[list[str | datetime]] = []
+        self._item_data: list[PasswordEntryData] = []
+
+        self._idx_lookup: dict[uuid.UUID, int] = {}
+        self._col_headers: list[str] = ["Title", "Username", "URL", "Created At"]
+
+    def rowCount(self, /, parent: QModelIndex = None):
+        return len(self._item_data)
+
+    def columnCount(self, /, parent: QModelIndex = None):
+        return len(self._col_headers)
+
+    def data(self, index, /, role):
+        if role == Qt.ItemDataRole.UserRole:
+            return self._item_data[index.row()]
+
+        if role == Qt.ItemDataRole.DisplayRole:
+            value = self._display_data[index.row()][index.column()]
+            if isinstance(value, datetime):
+                return value.strftime("%x %X")
+
+            if isinstance(value, AnyUrl):
+                return str(value)
+
+            if value is None:
+                logger.debug("Received 'None' value at cell [%d, %d]", index.row(), index.column())
+                return ""
+
+            return value
+
+        return None
+
+    def headerData(self, section, orientation, /, role):
+        if role != Qt.ItemDataRole.DisplayRole:
+            return None
+
+        if orientation == Qt.Orientation.Horizontal:
+            return self._col_headers[section]
+
+        if orientation == Qt.Orientation.Vertical:
+            return super().headerData(section, orientation, role)
+
+    def load_entries(self, entries: list[PasswordEntryData]):
+        self.beginResetModel()
+
+        self._display_data.clear()
+        self._item_data.clear()
+
+        self._idx_lookup.clear()
+        logger.debug("Cleared all password entries")
+
+        for entry in entries:
+            display_data = [entry.title, entry.username, entry.url, entry.created_at]
+            self._display_data.append(display_data)
+
+            self._item_data.append(entry)
+            self._idx_lookup[entry.entry_id] = self._item_data.index(entry)
+
+        logger.debug("Added %d entries to password entry model", len(self._item_data))
+        self.endResetModel()
+
+    def add_entry(self, entry: PasswordEntryData):
+        display_data = [entry.title, entry.username, entry.url, entry.created_at]
+
+        self._item_data.append(entry)
+        self._display_data.append(display_data)
+
+        self._idx_lookup[entry.entry_id] = self._item_data.index(entry)
+        self.layoutChanged.emit()
+
+    def delete_entry(self, data: PasswordEntryData):
+        row = self._idx_lookup.get(data.entry_id, None)
+        if row is None:
+            logger.warning("Entry ID '%s' not in model", data.entry_id)
+            return
+
+        del self._item_data[row]
+        del self._display_data[row]
+
+        del self._idx_lookup[data.entry_id]
+        self.layoutChanged.emit()
+
+    def update_entry(self, data: PasswordEntryData):
+        row = self._idx_lookup.get(data.entry_id, None)
+        if row is None:
+            logger.warning("Entry ID '%s' not in model", data.entry_id)
+            return
+
+        display_data = [data.title, data.username, data.url, data.created_at]
+
+        self._item_data[row] = data
+        self._display_data[row] = display_data
+
+        top = self.index(row, 0)
+        bottom = self.index(row, self.columnCount() - 1)
+
+        self.dataChanged.emit(top, bottom)
+
+
+class RecentDatabasesListModel(QAbstractListModel):
+    def __init__(self, /, parent: "DatabasesTabController" = None):
+        super().__init__(parent)
+        self.db_ctrl = parent
+
+        self._display_data: list[str] = []
+        self._item_data: list[Path] = []
+
+    def rowCount(self, /, parent: QModelIndex = None):
+        return len(self._item_data)
+
+    def data(self, index, /, role):
+        if role == Qt.ItemDataRole.UserRole:
+            return self._item_data[index.row()]
+
+        if role == Qt.ItemDataRole.DisplayRole:
+            value = self._display_data[index.row()]
+            return value
+
+        return None
+
+    def populate_model(self, recent_databases: list[Path]):
+        self.beginResetModel()
+
+        self._display_data.clear()
+        self._item_data.clear()
+
+        logger.debug("Cleared all model entries")
+
+        for path in recent_databases:
+            self._item_data.append(path)
+            self._display_data.append(str(path))
+
+        logger.debug("Added %d entries to model", len(self._item_data))
+        self.endResetModel()