+++ CHANGES.md
@@ -1,24 +1,36 @@
-# Implement most of server sync functionality
+# Clean up passwords tab code
 
 **Version**: v0.1.0
 
-**Date:** 17/07/2025
+**Date:** 19/07/2025
 
 ## Additions
 
-**`/app/models.py`**:
+**`/app/models/`**:
 
-* Added `entry_id` to `EditedEntryWithGroup` as it's the simplest way to keep track of the existing entry.
+* Added folder to store PySide6 models and state models.
 
-**`/app/serversync/`**:
+**`/app/serversync/client.py`**:
 
-* Added abstraction to the generated client as a simpler way to sync with the server.
+* Added option to set client to be enabled/disabled when initializing.
+
+**`/app/controllers/entry_data.py`**:
+
+* Added as the helper to manage data within the passwords tab.
 
 **`/app/controllers/tabs/passwords.py`**:
 
 * Added a way to optionally retrieve a loaded client.
 
 ## Changes
+
+**`/app/models.py`**:
+
+* Moved file to `/app/models/` directory.
+
+**`/app/controllers/sync_client.py`**:
+
+* No longer emits `loadWithoutSync`, instead only emits `clientLoaded` that has `self.enabled` set.
 
 **`/app/controllers/apps.py`**:
 
@@ -27,9 +39,21 @@
 
 **`/app/controllers/tabs/databases.py`**:
 
-* Many changes to how it handles the local database and server sync, needs a lot of cleanup.
+* Moved list model to `/app/models/ui.py`.
+
+**`/app/controllers/tabs/setting.py`**:
+
+* Unlocking the database settings tab now happens when client is loaded to let everything load first.
+* Now locks the edit sync info button while waiting for a request to complete.
+
+**`/app/controllers/tabs/passwords.py`**:
+
+* Moved table model to `/app/models/ui.py`.
+* Cleaned up main controller, setup now happens only after both database and client have been loaded.
+* Entries and groups controllers now require the database and client instances on initialization instead of taking the objects from the parent.
+* Entries controller no longer depends on a `QModelIndex` and instead the UUID/model.
 
 ## Misc
 
-* Will need a lot of cleanup to make sure the sync code doesn't break or become hard to maintain,
-  this will also make it easier to implement new features.
+* Cleanup of database and server sync code is halfway done, next is to clean up the groups and use
+  a real tree model instead of relying on a standard item model.